company policy for superadminand admin

+++ app/Http/Controllers/CompanyController.php
@@ -1,38 +1,36 @@
 <?php
 
-namespace App\Http\Controllers\SuperAdmin;
+namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCompanyRequest;
 use App\Models\Company;
+use App\Models\Role;
 use App\Models\TemporaryFile;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 class CompanyController extends Controller
 {
     /**
-     * Display a search companies.
-     */
-    public function search()
-    {
-        $companies = Company::all();
-        return view('company.index', compact('companies'));
-    }
-
-    /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny', Company::class);
 
-        if (request('search')) {
-            $companies = Company::where('nama', 'LIKE', '%' . request('search') . '%')->paginate(5);
-        } else {
-            $companies = Company::paginate(5);
+        if(auth()->user()->role_id === Role::$IS_SUPERADMIN) {
+            if (request('search')) {
+                $companies = Company::where('nama', 'LIKE', '%' . request('search') . '%')->paginate(5);
+            } else {
+                $companies = Company::paginate(5);
+            }
+            return view('superadmin.company.index', compact('companies'));
         }
-        return view('company.index', compact('companies'));
+
+        $companies = Company::where(['is_public' => 1])->where('nama', 'LIKE', '%' . request('search') . '%')->paginate(5);
+        $belongsTo = auth()->user()->company;
+        return view('admin.company.index', compact('companies', 'belongsTo'));
     }
 
     /**
@@ -40,7 +38,9 @@ class CompanyController extends Controller
      */
     public function create()
     {
-        return view('company.create');
+        $this->authorize('create', Company::class);
+
+        return view('superadmin.company.create');
     }
 
     /**
@@ -48,30 +48,34 @@ class CompanyController extends Controller
      */
     public function store(StoreCompanyRequest $request)
     {
+        $this->authorize('create', Company::class);
+
         // validate request
         $validated = $request->validated();
         $tempFile = TemporaryFile::where('folder', $request->img)->first();
-        $filename = uniqid() . '-' . $tempFile->filename;
-
-        // Store the image at the specified path.
-        FIle::copy(
-            storage_path("app\\img\\tmp\\$tempFile->folder\\$tempFile->filename"),
-            storage_path("app\\public\\company\\$filename")
-        );
-
-        // Get the logo file name.
-        $validated['logo'] = $filename;
-
-        // Create a project with the validated data.
-        Company::create($validated);
-
+        dd($request);
+        // dd(TemporaryFile::where(['folder'=>$request->img])->get());
         if ($tempFile) {
+            $filename = uniqid() . '-' . $tempFile->filename;
             File::deleteDirectory(storage_path("app\\img\\tmp\\$tempFile->folder"));
-        }
-        $tempFile->delete();
+            $tempFile->delete();
 
-        // Redirect the user to the project list.
-        return redirect()->route('company.index')->with('success', "Data Company $request->nama Berhasil Dibuat");
+            // Store the image at the specified path.
+            File::copy(
+                storage_path("app\\img\\tmp\\$tempFile->folder\\$tempFile->filename"),
+                storage_path("app\\public\\company\\$filename")
+            );
+
+            // Get the logo file name.
+            $validated['logo'] = $filename;
+
+            // Create a project with the validated data.
+            Company::create($validated);
+            // Redirect the user to the project list.
+            return redirect()->route('company.index')->with('success', "Data Company $request->nama Berhasil Dibuat");
+        }
+
+        return redirect()->route('company.index')->with('failed', "Data Company $request->nama Gagal Dibuat");
     }
 
     /**
@@ -79,7 +83,9 @@ class CompanyController extends Controller
      */
     public function show(Company $company)
     {
-        return view('company.show', compact('company'));
+        $this->authorize('view', $company);
+
+        return view('superadmin.company.show', compact('company'));
     }
 
     /**
@@ -87,7 +93,10 @@ class CompanyController extends Controller
      */
     public function edit(Company $company)
     {
-        return view('company.edit', compact('company'));
+        $this->authorize('update', $company);
+
+
+        return view('superadmin.company.edit', compact('company'));
     }
 
     /**
@@ -95,6 +104,7 @@ class CompanyController extends Controller
      */
     public function update(StoreCompanyRequest $request, Company $company)
     {
+        $this->authorize('update', $company);
         // validate request
         $validated = $request->validated();
 
@@ -130,6 +140,7 @@ class CompanyController extends Controller
      */
     public function destroy(Company $company)
     {
+        $this->authorize('delete', $company);
         $success = $company->deleteOrFail();
 
         if ($success) return redirect()->route('company.index')->with('success', 'Data Berhasil Dihapus');
